Bill OCR/Text Extract Part/requirement.txt
google.generativeai
Pillow